format changes

+++ src/java/peregrine/sysstat/DiskStat.java
@@ -138,7 +138,8 @@ public class DiskStat extends BaseStat implements Diffable<DiskStat> {
         StringBuilder buff = new StringBuilder();
 
         buff.append( String.format( "%10s %,15d %,15d %,15d %,15d %,15.2f %15.2f",
-                                    name, reads, writes, readBytes.longValue(), writtenBytes.longValue(), avg_req_size, util ) );
+                                    name, reads.intValue(), writes.intValue(),
+                                    readBytes.longValue(), writtenBytes.longValue(), avg_req_size, util ) );
 
         return buff.toString();
